Found a bug in the temperature sensor which had a uint8_t return value while the function was declared to return a uint32_t. Manually changed the data type to fix. Also noticed Pressure Sensor and Magnetometer didn't initialize correctly and trying forever in a do-while loop fixed the problem.

+++ WeatherStation/Drivers/BSP/B-L475E-IOT01/stm32l475e_iot01_tsensor.c
@@ -50,7 +50,7 @@ static TSENSOR_DrvTypeDef *tsensor_drv;
   */
 uint32_t BSP_TSENSOR_Init(void)
 {  
-  uint8_t ret = TSENSOR_ERROR;
+  uint32_t ret = TSENSOR_ERROR;
 
 #ifdef USE_LPS22HB_TEMP
   tsensor_drv = &LPS22HB_T_Drv;
